compilation fixes

+++ src/main/scala/learning/FeaturizationDefaults.scala
@@ -16,12 +16,12 @@ object BaseFeaturization {
   * Utility class of default featurizations
   */
 object FeaturizationDefaults {
-  def planFeaturization(input : QueryInstruction) : (ArrayBuffer[Int], ArrayBuffer[Int], ArrayBuffer[Int]) = {
+  def planFeaturization(input : QueryInstruction) : (ArrayBuffer[Double], ArrayBuffer[Double], ArrayBuffer[Double]) = {
     // we cannot featurize the relations - they vary in size and may be tremendous in size.
     // we *can* featurize the cardinality.
-    var featureVector = new ArrayBuffer[Int]()
-    var relationVector = new ArrayBuffer[Int]()
-    var constraintVector = new ArrayBuffer[Int]()
+    var featureVector = new ArrayBuffer[Double]()
+    var relationVector = new ArrayBuffer[Double]()
+    var constraintVector = new ArrayBuffer[Double]()
     for (relation <- input.relations) {
       if (relation.isLeft) {
         relationVector += relation.left.get.relationContent.size
@@ -47,7 +47,7 @@ object FeaturizationDefaults {
   // use isCommutative flag
   // use isAssociative?
   /* Global list is a parameter in case we want to do a one-hot featurization */
-  def joinFeaturization(r1 : RelationStub, r2 : RelationStub, globalRelations : Vector[String]): (ArrayBuffer[Int], ArrayBuffer[Int]) = {
+  def joinFeaturization(r1 : RelationStub, r2 : RelationStub, globalRelations : Vector[String]): (ArrayBuffer[Double], ArrayBuffer[Double]) = {
     val dummyR1 = new Join(ArrayBuffer[Either[RelationStub, QueryInstruction]](Left(r1)), ArrayBuffer[ConstraintStub]())
     val r1Features = planFeaturization(dummyR1)._2
     val dummyR2 = new Join(ArrayBuffer[Either[RelationStub, QueryInstruction]](Left(r2)), ArrayBuffer[ConstraintStub]())
